Extract screens into separate components
