+++ app/src/main/java/com/example/pgpandy/InboxScreen.kt
@@ -1,0 +1,15 @@
+package com.example.pgpandy
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+
+@Composable
+fun InboxScreen() {
+    Box(modifier = Modifier.fillMaxSize()) {
+        Text("Inbox", modifier = Modifier.align(Alignment.Center))
+    }
+}
