+++ app/src/main/java/com/example/pgpandy/KeyListScreen.kt
@@ -1,0 +1,31 @@
+package com.example.pgpandy
+
+import androidx.compose.material3.FloatingActionButton
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Key
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.graphics.Color
+
+@Composable
+fun KeyListScreen(onAddKey: () -> Unit) {
+    Box(modifier = Modifier.fillMaxSize()) {
+        Text("No private keys added.", modifier = Modifier.align(Alignment.Center))
+
+        FloatingActionButton(
+            onClick = onAddKey,
+            modifier = Modifier.align(Alignment.BottomEnd).padding(16.dp),
+            containerColor = Color(0xFF440020),
+            contentColor = Color.White
+        ) {
+            Icon(Icons.Default.Key, contentDescription = null)
+        }
+    }
+}
